URLForm component is stateless
but a little imperative as well.

Instead of keeping `valid` as a state property I use a feature of React 1.14 where a callback can be passed as `ref` prop value:

https://facebook.github.io/react/docs/more-about-refs.html#the-ref-callback-attribute

That way I can save reference to `input` and `button` elements in a closure and use them to validate `input.value` whenever it changes. The `onChange` callback enables or disables `button`. This part is imperative and it smells a little.